iFixitApi: send PSK
We've changed our security mechanism to allow usage of a PSK instead of
just "are you local". This allows us to not care where the service is
run, allowing dev and production to re-use the same docker-compose and
use the public DNS of ifixit and not worry about /etc/hosts.

The other end of this has already been set up.