Feature: Product Edit Page Redirector
Akeneo doesn't have a way to open the product edit page based on a
product's sku, or a product-model's code. Without this, the only way to
do it is via the cumbersome UI.

Here we add a new controller that only serves to redirect the browser
to the appropriate edit page of the given sku or productcode.

Note: we use one url, but route to different methods based on the length
of the parameter