PostSave: listen to events from AttributeOptions
When you edit just one of the options of an akeneo simple select
attribute, the only post-save events you get are for the Option,
not the Attribute itself.

So, here we start listening to these events and extracting the relevant
info. Our app handles refreshing the info for an entire attribute
so we can send the same type of "attribute changed" notification.