Price Updates: Use custom comparator
The comparator ends up being used earlier in the chain and has access to
both original prices and the ewn prices in a closer-to-original form.

Here we can implement the logic as we need to support both CSVs and the
UI.

Specifically, the UI produces an exhaustive list of currency amounts
cause it has one UI element per input (nulls for un-filled-in-fields).
While CSV imports only produce one currency for each column in the CSV.

We want to effectively treat the incoming data as a patch against
the original values. Nulls remove currencies and non-nulls replace the
values.

The is just a little tricky cause the incoming data isn't a map, but an
array of [currency, amount] tuples. So, we convert to a currency =>
amount map first for easy merging.

Previously:
The previous behavior treated the new values as exhaustive and thus a
single-currency column from the CSV ended up replacing the whole
collection of [currency, amount] tuples with the single one.