PricesComparator: drop thousands separator
Including the thousands separator in prices that were sent to the
next level ended up failing saves that included a price > 1000 (thus
including a separator).

This class was copied from the akeneo one and slightly modified. In
Akeneo, these values are only used for comparison and thrown away,
accepting the input wholesale or not at all. In our version, they are
used for comparison and value merging, which is why the later stages are
complaining about the formatting.