Price uploads: override some relevant services
Problem: We want to be able to unset (set to null) individual prices via
csv uploads. There are several services involved so we have to customize
them to achieve this.

First step is to load our custom versions (and not yet change their
behavior). We used different techniques to override the services where
appropriate, including decoration (activating the service with
higher-priority tags) and service replacement by extending the class
(just redefining the service in the yaml).