Price Uploads: support nulls
Change two classes to be able to support uploading prices with blank
price columns (unsetting the value for that currency).

==== ColumnsMerger
This class takes CSV fields like "price-USD" and "price-CAD" and turns
into them into a string of comma-separated values like "12.99 USD, 14.95
CAD". It needs to *not* filter out nulls.

This means that a blank `price-USD` value going in will get converted
to: " USD, 14.95 CAD"

==== PriceCollectionValueFactory
Use array_key_exists() instead of isset() so that null (or blank)
amounts don't end up throwing an exception.