PostSave: record skus on in-between events
Commit bdcfa3812e3c3cd72e6d41a01a7491427c403013 desired to listen events
on ProductModels and extract skus from them. Turns out, it had a bug.

We were only trying to extract skus from the last post-save event in the
group. We want to send the skus_changed hook  on the last event, but we
want to extract skus from *all* events. The two lines were just in the
wrong place (inside the if()).

The result was that if you uploaded a CSV and modified Product Models,
we wouldn't send any skus_changed notifications. This was because the
last `post-save` was never at level 0 (it was always enclosed by a
`post-save-all`)