ifixit.com api requests: only retry non-200
Wow. Oops. Turns out this middleware relies on us determining if we
should retry or not including examining the status code. There are no
docs and I then I misread the source code (or saw simplistic examples or
something).

We're submitting to the Akeneo Events api on ifixit.com and it always
gives back 200s, so let's consider anything non-200 as a failure.

This has been a bug since 2022-09-07:
https://github.com/iFixit/pim-community-standard/pull/20j