composer: add manual post-upgrade script
There's a few things we need to do after running this script. Since it's
run any time we hit composer-install, it was a bit much and it was easy
to forget to do the following bits. So, let's make this script manual
(giving it a custom name) and put our follow-up bits in there.

In specific, we need to apply a patch to a migration and we need to undo
the overwriting of the .env file by the upgrade script.

Also, include running this in the "upgrade" Make target. Reformat it
too, I'm not sure why we invoke make a bunch of times...