Job-consumer container: run as fedora user
Other services end up running as appropriate users for their service,
but since our docker image doesn't force a user, it was defaulting to
root.

This meant that if this process created any dirs or files, they'd only
be accessible to root. This was breaking some features that allowed
downloading the results or logs of background jobs.