skus_changed events: pass a priority adjustment
We want to prioritize single-product edits made from the web UI
over batch edits made via CSV. One happens in FPM and the other in CLI.

We can send through an adjustment in the request body indicating how to
tweak the priority.

I tried to find a way to differentiate CSV edits with the Batch edit UI
on the web but couldn't. I even tried getting the stack trace, but the
traces were identical as far as I could tell. So, for now, they are
lumped together as lower priority.